+++ paddles/deque.py
@@ -82,7 +82,7 @@ class LinkedListDeque:
         The members are added to the deque in the order they are in `sequence`.
         To create an empty deque, call `LinkedListDeque()` or `LinkedListDeque([])`.
 
-        Complexity: O(len(`sequence`))
+        Complexity: O(n), with n = `len(sequence)`
         """
         self._head = None
         self._tail = None
@@ -96,7 +96,7 @@ class LinkedListDeque:
 
         The string is 'LinkedListDeque([front member, ..., back member])'.
 
-        Complexity: O(self.size())
+        Complexity: O(n), with n = `self.size()`
         """
         strings = []
         current = self._head
@@ -113,7 +113,7 @@ class LinkedListDeque:
         return self._length
 
     def front(self) -> Any:
-        """Return the item at the front of the deque, without removing it.
+        """Return the member at the front of the deque, without removing it.
 
         Raise `ValueError` if the deque is empty.
 
@@ -125,7 +125,7 @@ class LinkedListDeque:
         return self._head[DATA]
 
     def back(self) -> Any:
-        """Return the item at the back of the deque, without removing it.
+        """Return the member at the back of the deque, without removing it.
 
         Raise `ValueError` if the deque is empty.
 
@@ -163,7 +163,7 @@ class LinkedListDeque:
         self._length += 1
 
     def take_front(self) -> Any:
-        """Remove and return the item at the front of the deque.
+        """Remove and return the member at the front of the deque.
 
         Raise `ValueError` if the deque is empty.
 
@@ -182,7 +182,7 @@ class LinkedListDeque:
         return item
 
     def take_back(self) -> Any:
-        """Remove and return the item at the back of the deque.
+        """Remove and return the member at the back of the deque.
 
         Raise `ValueError` if the deque is empty.